+++ Runtime/DB/GameProcess DB/GameProcessDB.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using OcDialogue.DB;
+using OcDialogue.DB.GameProcess_DB;
 using OcUtility;
 using Sirenix.OdinInspector;
 using UnityEditor;
@@ -14,11 +15,15 @@ namespace OcDialogue.DB
     public class GameProcessDB : OcData, IDataRowUser
     {
         public static GameProcessDB Instance => DBManager.Instance.GameProcessDB;
+        public IGamePlayer GamePlayer { get; set; }
+        public DateTime GameTime { get; set; }
+        public TimeSpan PlayTimeUntilLastSession { get; set; }
+        public TimeSpan PlayTimeSinceThisSession { get; set; }
         public override string Address => "GameProcessData";
         public DataRowContainer DataRowContainer => dataRowContainer;
+        
         public bool usePreset;
         public DataRowContainer dataRowContainer;
-        public List<DynamicData> DynamicDataList;
 
         public event Action OnRuntimeValueChanged;
         public void Init()
@@ -43,7 +48,8 @@ namespace OcDialogue.DB
         public void Overwrite(GameProcessSaveData saveData)
         {
             dataRowContainer.Overwrite(saveData.DataRowContainerDict);
-            DynamicDataList = saveData.DynamicDataList;
+            GameTime = saveData.GameTime;
+            PlayTimeUntilLastSession = saveData.PlayTime;
         }
 
         public GameProcessSaveData GetSaveData()
@@ -51,48 +57,12 @@ namespace OcDialogue.DB
             var data = new GameProcessSaveData()
             {
                 DataRowContainerDict = dataRowContainer.GetSaveData(),
-                DynamicDataList = new List<DynamicData>(DynamicDataList)
+                PlayerTransform = GamePlayer.LastSafeTransformData,
+                PlayerHP = GamePlayer.HP,
+                GameTime = GameTime,
+                PlayTime = PlayTimeUntilLastSession + PlayTimeSinceThisSession
             };
             return data;
-        }
-
-        public void ReadDynamicDataOrRegister(DynamicDataBehaviour dynamicDataBehaviour)
-        {
-            var data = DynamicDataList.FirstOrDefault(x => x.Key == dynamicDataBehaviour.Key);
-            if (data == null)
-            {
-                DynamicDataList.Add(new DynamicData(dynamicDataBehaviour));
-            }
-            else
-            {
-                if (dynamicDataBehaviour.UseTransformData)
-                {
-                    var transform = dynamicDataBehaviour.transform;
-                    transform.position = data.TransformData.position;
-                    transform.rotation = data.TransformData.rotation;
-                    transform.localScale = data.TransformData.scale;
-                }
-                dynamicDataBehaviour.PrimitiveData = data.PrimitiveData;
-            }
-        }
-
-        public void UpdateDynamicData(DynamicDataBehaviour dynamicDataBehaviour)
-        {
-            var data = DynamicDataList.FirstOrDefault(x => x.Key == dynamicDataBehaviour.Key);
-            if (data == null)
-            {
-                DynamicDataList.Add(new DynamicData(dynamicDataBehaviour));
-            }
-            else
-            {
-                if (dynamicDataBehaviour.UseTransformData)
-                {
-                    data.TransformData = new TransformData(dynamicDataBehaviour.transform);
-                }
-
-                data.PrimitiveData = dynamicDataBehaviour.PrimitiveData;
-            }
-            OnRuntimeValueChanged?.Invoke();
         }
 
 #if UNITY_EDITOR