+++ Runtime/DB/NPC DB/INPCCharacter.cs
@@ -1,0 +1,11 @@
+using UnityEngine;
+
+namespace OcDialogue.DB
+{
+    public interface INPCCharacter
+    {
+        GameObject gameObject { get; }
+        Transform transform { get; }
+        NPC NPC { get; }
+    }
+}